requirements.txt
telethon
openai
python-dotenv
asyncio
